+++ src/fraudtwin/simulation/__init__.py
@@ -2,7 +2,11 @@
 
 from fraudtwin.simulation.behavior import BehaviorDataset, BehaviorGenerator, generate_behavior
 from fraudtwin.simulation.generator import EntityDataset, EntityGenerator, generate_entities
-from fraudtwin.simulation.payments import PaymentDataset, PaymentGenerator
+from fraudtwin.simulation.payments import (
+    PaymentDataset,
+    PaymentGenerator,
+    count_pix_lifecycle_events,
+)
 
 __all__ = [
     "BehaviorDataset",
@@ -11,6 +15,7 @@ __all__ = [
     "EntityGenerator",
     "PaymentDataset",
     "PaymentGenerator",
+    "count_pix_lifecycle_events",
     "generate_behavior",
     "generate_entities",
 ]